restructure mock and add pull request
