+++ OnToology/mock/fork.py
@@ -1,0 +1,118 @@
+
+
+
+
+def get_fork_dict(repo):
+    user, name = repo.split("/")
+    fork = {
+  "id": 1296269,
+  "node_id": "MDEwOlJlcG9zaXRvcnkxMjk2MjY5",
+  "name": name,
+  "full_name": repo,
+  "owner": {
+    "login": user,
+    "id": 1,
+    "node_id": "MDQ6VXNlcjE=",
+    "avatar_url": "https://github.com/images/error/%s_happy.gif" % user,
+    "gravatar_id": "",
+    "url": "https://api.github.com/users/%s"% user,
+    "html_url": "https://github.com/%s"% user,
+    "followers_url": "https://api.github.com/users/%s/followers"% user,
+    "following_url": "https://api.github.com/users/%s/following{/other_user}"% user,
+    "gists_url": "https://api.github.com/users/%s/gists{/gist_id}"% user,
+    "starred_url": "https://api.github.com/users/%s/starred{/owner}{/repo}"% user,
+    "subscriptions_url": "https://api.github.com/users/%s/subscriptions"% user,
+    "organizations_url": "https://api.github.com/users/%s/orgs"% user,
+    "repos_url": "https://api.github.com/users/%s/repos"% user,
+    "events_url": "https://api.github.com/users/%s/events{/privacy}"% user,
+    "received_events_url": "https://api.github.com/users/%s/received_events"% user,
+    "type": "User",
+    "site_admin": False
+  },
+  "private": False,
+  "html_url": "https://github.com/%s" % repo,
+  "description": "This your first repo!",
+  "fork": True,
+  "url": "https://api.github.com/repos/%s"% repo,
+  "archive_url": "http://api.github.com/repos/%s/{archive_format}{/ref}"% repo,
+  "assignees_url": "http://api.github.com/repos/%s/assignees{/user}"% repo,
+  "blobs_url": "http://api.github.com/repos/%s/git/blobs{/sha}"% repo,
+  "branches_url": "http://api.github.com/repos/%s/branches{/branch}"% repo,
+  "collaborators_url": "http://api.github.com/repos/%s/collaborators{/collaborator}"% repo,
+  "comments_url": "http://api.github.com/repos/%s/comments{/number}"% repo,
+  "commits_url": "http://api.github.com/repos/%s/commits{/sha}"% repo,
+  "compare_url": "http://api.github.com/repos/%s/compare/{base}...{head}"% repo,
+  "contents_url": "http://api.github.com/repos/%s/contents/{+path}"% repo,
+  "contributors_url": "http://api.github.com/repos/%s/contributors"% repo,
+  "deployments_url": "http://api.github.com/repos/%s/deployments"% repo,
+  "downloads_url": "http://api.github.com/repos/%s/downloads"% repo,
+  "events_url": "http://api.github.com/repos/%s/events"% repo,
+  "forks_url": "http://api.github.com/repos/%s/forks"% repo,
+  "git_commits_url": "http://api.github.com/repos/%s/git/commits{/sha}"% repo,
+  "git_refs_url": "http://api.github.com/repos/%s/git/refs{/sha}"% repo,
+  "git_tags_url": "http://api.github.com/repos/%s/git/tags{/sha}"% repo,
+  "git_url": "git:github.com/%s.git"% repo,
+  "issue_comment_url": "http://api.github.com/repos/%s/issues/comments{/number}"% repo,
+  "issue_events_url": "http://api.github.com/repos/%s/issues/events{/number}"% repo,
+  "issues_url": "http://api.github.com/repos/%s/issues{/number}"% repo,
+  "keys_url": "http://api.github.com/repos/%s/keys{/key_id}"% repo,
+  "labels_url": "http://api.github.com/repos/%s/labels{/name}"% repo,
+  "languages_url": "http://api.github.com/repos/%s/languages"% repo,
+  "merges_url": "http://api.github.com/repos/%s/merges"% repo,
+  "milestones_url": "http://api.github.com/repos/%s/milestones{/number}"% repo,
+  "notifications_url": "http://api.github.com/repos/%s/notifications{?since,all,participating}"% repo,
+  "pulls_url": "http://api.github.com/repos/%s/pulls{/number}"% repo,
+  "releases_url": "http://api.github.com/repos/%s/releases{/id}"% repo,
+  "ssh_url": "git@github.com:%s.git"% repo,
+  "stargazers_url": "http://api.github.com/repos/%s/stargazers"% repo,
+  "statuses_url": "http://api.github.com/repos/%s/statuses/{sha}"% repo,
+  "subscribers_url": "http://api.github.com/repos/%s/subscribers"% repo,
+  "subscription_url": "http://api.github.com/repos/%s/subscription"% repo,
+  "tags_url": "http://api.github.com/repos/%s/tags"% repo,
+  "teams_url": "http://api.github.com/repos/%s/teams"% repo,
+  "trees_url": "http://api.github.com/repos/%s/git/trees{/sha}"% repo,
+  "clone_url": "https://github.com/%s.git"% repo,
+  "mirror_url": "git:git.example.com/%s"% repo,
+  "hooks_url": "http://api.github.com/repos/%s/hooks"% repo,
+  "svn_url": "https://svn.github.com/%s"% repo,
+  "homepage": "https://github.com",
+  "language": None,
+  "forks_count": 9,
+  "stargazers_count": 80,
+  "watchers_count": 80,
+  "size": 108,
+  "default_branch": "master",
+  "open_issues_count": 0,
+  "is_template": True,
+  "topics": [
+    "abc",
+    "atom",
+    "electron",
+    "api"
+  ],
+  "has_issues": True,
+  "has_projects": True,
+  "has_wiki": True,
+  "has_pages": False,
+  "has_downloads": True,
+  "archived": False,
+  "disabled": False,
+  "visibility": "public",
+  "pushed_at": "2011-01-26T19:06:43Z",
+  "created_at": "2011-01-26T19:01:12Z",
+  "updated_at": "2011-01-26T19:14:43Z",
+  "permissions": {
+    "admin": False,
+    "push": False,
+    "pull": True
+  },
+  "allow_rebase_merge": True,
+  "template_repository": None,
+  "temp_clone_token": "ABTLWHOULUVAXGTRYU7OC2876QJ2O",
+  "allow_squash_merge": True,
+  "delete_branch_on_merge": True,
+  "allow_merge_commit": True,
+  "subscribers_count": 42,
+  "network_count": 0
+}
+    return fork
