+++ OnToology/mock/user.py
@@ -1,0 +1,88 @@
+
+
+def get_auth_user_dict(user):
+    a_user_dict = {
+        "login": user,
+        "id": 1,
+        "node_id": "MDQ6VXNlcjE=",
+        "avatar_url": "https://github.com/images/error/%s_happy.gif" % user,
+        "gravatar_id": "",
+        "url": "https://api.github.com/users/%s" % user,
+        "html_url": "https://github.com/%s" % user,
+        "followers_url": "https://api.github.com/users/%s/followers" % user,
+        "following_url": "https://api.github.com/users/%s/following{/other_user}" % user,
+        "gists_url": "https://api.github.com/users/%s/gists{/gist_id}" % user,
+        "starred_url": "https://api.github.com/users/%s/starred{/owner}{/repo}" % user,
+        "subscriptions_url": "https://api.github.com/users/%s/subscriptions" % user,
+        "organizations_url": "https://api.github.com/users/%s/orgs" % user,
+        "repos_url": "https://api.github.com/users/%s/repos" % user,
+        "events_url": "https://api.github.com/users/%s/events{/privacy}" % user,
+        "received_events_url": "https://api.github.com/users/%s/received_events" % user,
+        "type": "User",
+        "site_admin": False,
+        "name": "monalisa %s" % user,
+        "company": "GitHub",
+        "blog": "https://github.com/blog",
+        "location": "San Francisco",
+        "email": "%s@github.com" % user,
+        "hireable": False,
+        "bio": "There once was...",
+        "twitter_username": "%s" % user,
+        "public_repos": 2,
+        "public_gists": 1,
+        "followers": 20,
+        "following": 0,
+        "created_at": "2008-01-14T04:33:35Z",
+        "updated_at": "2008-01-14T04:33:35Z",
+        "private_gists": 81,
+        "total_private_repos": 100,
+        "owned_private_repos": 100,
+        "disk_usage": 10000,
+        "collaborators": 8,
+        "two_factor_authentication": True,
+        "plan": {
+            "name": "Medium",
+            "space": 400,
+            "private_repos": 20,
+            "collaborators": 0
+        }
+    }
+    return a_user_dict
+
+
+def get_user_dict(user):
+    user_dict = {
+        "login": user,
+        "id": 1,
+        "node_id": "MDQ6VXNlcjE=",
+        "avatar_url": "https://github.com/images/error/%s_happy.gif" % user,
+        "gravatar_id": "",
+        "url": "https://api.github.com/users/%s" % user,
+        "html_url": "https://github.com/%s" % user,
+        "followers_url": "https://api.github.com/users/%s/followers" % user,
+        "following_url": "https://api.github.com/users/%s/following{/other_user}" % user,
+        "gists_url": "https://api.github.com/users/%s/gists{/gist_id}" % user,
+        "starred_url": "https://api.github.com/users/%s/starred{/owner}{/repo}" % user,
+        "subscriptions_url": "https://api.github.com/users/%s/subscriptions" % user,
+        "organizations_url": "https://api.github.com/users/%s/orgs" % user,
+        "repos_url": "https://api.github.com/users/%s/repos" % user,
+        "events_url": "https://api.github.com/users/%s/events{/privacy}" % user,
+        "received_events_url": "https://api.github.com/users/%s/received_events" % user,
+        "type": "User",
+        "site_admin": False,
+        "name": "monalisa ahmad88me",
+        "company": "GitHub",
+        "blog": "https://github.com/blog",
+        "location": "San Francisco",
+        "email": "%s@gmail.com" % user,
+        "hireable": False,
+        "bio": "There once was...",
+        "twitter_username": "%s" % user,
+        "public_repos": 2,
+        "public_gists": 1,
+        "followers": 20,
+        "following": 0,
+        "created_at": "2008-01-14T04:33:35Z",
+        "updated_at": "2008-01-14T04:33:35Z",
+    }
+    return user_dict
